Move extensions to separate file
